highlight search query in 'Open Quickly' results
